+++ internal/interpreter/function_statements.go
@@ -1,0 +1,35 @@
+package interpreter
+
+import "github.com/formancehq/numscript/internal/parser"
+
+func setTxMeta(st *programState, r parser.Range, args []Value) InterpreterError {
+	p := NewArgsParser(args)
+	key := parseArg(p, r, expectString)
+	meta := parseArg(p, r, expectAnything)
+	err := p.parse()
+	if err != nil {
+		return err
+	}
+
+	st.TxMeta[*key] = *meta
+	return nil
+}
+
+func setAccountMeta(st *programState, r parser.Range, args []Value) InterpreterError {
+	p := NewArgsParser(args)
+	account := parseArg(p, r, expectAccount)
+	key := parseArg(p, r, expectString)
+	meta := parseArg(p, r, expectAnything)
+	err := p.parse()
+	if err != nil {
+		return err
+	}
+
+	accountMeta := defaultMapGet(st.SetAccountsMeta, *account, func() AccountMetadata {
+		return AccountMetadata{}
+	})
+
+	accountMeta[*key] = (*meta).String()
+
+	return nil
+}
